requirements.txt
fastapi[standard]==0.115.5
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.0
pandas==2.1.4
psycopg2-binary==2.9.6
xgboost==1.7.6
sckit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
sys
datetime